Search for executables in PATH
Added a mechanism in python code that searches for executables in the
whole PATH locations, if they are not in src/shorah/bin. This works for
installations obtained with pip install, and should also work with
autotools.

Minor modifications include proper logging arguments and use of double
quotes in docstrings (not yet PEP257 compliant).